# specs/001-heliolus-platform/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Heliolus Compliance Assessment Platform API
  description: API for compliance assessment, gap analysis, and vendor marketplace
  version: 1.0.0
  contact:
    name: Heliolus Support
    email: support@heliolus.com
servers:
  - url: https://api.heliolus.com/v1
    description: Production
  - url: https://staging-api.heliolus.com/v1
    description: Staging
  - url: http://localhost:3001/v1
    description: Development
security:
  - bearerAuth: []
...
paths:
...
  /organizations:
    post:
      tags: [Organizations]
      summary: Create organization
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, country]
              properties:
                name:
                  type: string
                website:
                  type: string
                  format: uri
                industry:
                  type: string
                size:
                  $ref: '#/components/schemas/CompanySize'
                country:
                  type: string
                region:
                  type: string
                description:
                  type: string
      responses:
        201:
          description: Organization created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        400:
          $ref: '#/components/responses/BadRequest'
...
  /templates:
    get:
      tags: [Templates]
      summary: List templates
      operationId: listTemplates
      parameters:
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/TemplateCategory'
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        200:
          description: Template list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Template'
  /templates/{id}:
    get:
      tags: [Templates]
      summary: Get template
      operationId: getTemplate
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: Template details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        404:
          $ref: '#/components/responses/NotFound'

  # Assessments
  /assessments:
    get:
      tags: [Assessments]
      summary: List assessments
      operationId: listAssessments
      parameters:
        - name: organizationId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AssessmentStatus'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        200:
          description: Assessment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Assessment'

    post:
      tags: [Assessments]
      summary: Create assessment
      operationId: createAssessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [organizationId, templateId]
              properties:
                organizationId:
                  type: string
                templateId:
                  type: string
      responses:
        201:
          description: Assessment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assessment'
        400:
          $ref: '#/components/responses/BadRequest'
        402:
          description: Insufficient credits
...
  /vendors:
    get:
      tags: [Vendors]
      summary: List vendors
      operationId: listVendors
      parameters:
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/VendorCategory'
        - name: featured
          in: query
          schema:
            type: boolean
        - name: verified
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        200:
          description: Vendor list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vendor'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
    offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
...
    Organization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        website:
          type: string
        industry:
          type: string
        size:
          $ref: '#/components/schemas/CompanySize'
        country:
          type: string
        region:
          type: string
        description:
          type: string
        onboardingCompleted:
          type: boolean
        riskProfile:
          $ref: '#/components/schemas/RiskProfile'
        createdAt:
          type: string
          format: date-time
    CompanySize:
      type: string
      enum: [STARTUP, SMB, MIDMARKET, ENTERPRISE]
    RiskProfile:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
...
    Template:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        category:
          $ref: '#/components/schemas/TemplateCategory'
        version:
          type: string
        isActive:
          type: boolean
        sections:
          type: array
          items:
            $ref: '#/components/schemas/TemplateSection'
    TemplateCategory:
      type: string
      enum: [FINANCIAL_CRIME, TRADE_COMPLIANCE, DATA_PRIVACY, CYBERSECURITY, ESG]
    TemplateSection:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        questions:
          type: array
          items:
            $ref: '#/components/schemas/TemplateQuestion'
        weight:
          type: number
        order:
          type: integer
    TemplateQuestion:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        type:
          $ref: '#/components/schemas/QuestionType'
        required:
          type: boolean
        options:
          type: array
          items:
            type: string
        helpText:
          type: string
        order:
          type: integer
    QuestionType:
      type: string
      enum: [TEXT, NUMBER, SELECT, MULTISELECT, BOOLEAN, FILE, DATE]
    Assessment:
      type: object
      properties:
        id:
          type: string
        organizationId:
          type: string
        templateId:
          type: string
        status:
          $ref: '#/components/schemas/AssessmentStatus'
        responses:
          type: object
        riskScore:
          type: number
        creditsUsed:
          type: integer
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    AssessmentStatus:
      type: string
      enum: [DRAFT, IN_PROGRESS, COMPLETED, FAILED]
...
    Vendor:
      type: object
      properties:
        id:
          type: string
        companyName:
          type: string
        website:
          type: string
        logo:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        categories:
          type: array
          items:
            $ref: '#/components/schemas/VendorCategory'
        featured:
          type: boolean
        verified:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
    VendorCategory:
      type: string
      enum:
        [
          KYC_AML,
          TRANSACTION_MONITORING,
          SANCTIONS_SCREENING,
          TRADE_SURVEILLANCE,
          RISK_ASSESSMENT,
          COMPLIANCE_TRAINING,
          REGULATORY_REPORTING,
          DATA_GOVERNANCE,
        ]
...